Fix: test 수정

+++ IrumOrder_BackEnd/src/test/java/Irumping/IrumOrder/service/CategoryServiceTest.java
@@ -23,7 +23,7 @@ class CategoryServiceTest {
         String categoryName = "testCategory";
 
         // When
-        Long categoryId = categoryService.createCategory(categoryName);
+        Integer categoryId = categoryService.createCategory(categoryName);
 
         // Then
         CategoryEntity category = categoryService.findCategoryById(categoryId);
